requirements.txt
python-dotenv
requests
pillow
numpy
tensorflow
streamlit
scikit-learn
pandas
nltk

